Create a new test file for ops queries
