+++ galley/formatted_queries.py
@@ -1,6 +1,7 @@
 import logging
 from re import M
 from typing import Dict, List, Optional
+
 from galley.enums import (
     DietaryFlagEnum,
     IngredientCategoryTagTypeEnum,
