+++ tests/test_queries.py
@@ -4,13 +4,12 @@ from unittest import TestCase, mock
 from galley.queries import (
     Query, get_menu_query, get_raw_menu_data,
     get_raw_recipes_data, recipe_connection_query,
-    get_ops_menu_query
 )
 from sgqlc.operation import Operation
 
 from tests.mock_responses import mock_recipes_data
 from tests.mock_responses.mock_menu_data import mock_menu
-from pprint import pprint
+
 
 logger = logging.getLogger(__name__)
 
@@ -619,120 +618,3 @@ class TestQueryGetRawRecipesData(TestCase):
         result = get_raw_recipes_data(['1', '2', '3'])
         self.assertEqual(mock_retrieval_method.call_count, 2)
         self.assertEqual(result, expected_recipe_data)
-
-
-class TestOpsMenuDataQuery(TestCase):
-    def setUp(self) -> None:
-        self.expected_query = '''query {
-            viewer {
-            menus(where: {date: ["2022-03-28"]}) {
-            id
-            name
-            date
-            location {
-            name
-            }
-            categoryValues {
-            id
-            name
-            category {
-            id
-            name
-            itemType
-            }
-            }
-            menuItems {
-            id
-            recipeId
-            categoryValues {
-            id
-            name
-            category {
-            id
-            name
-            itemType
-            }
-            }
-            recipe {
-            files {
-            photos {
-            sourceUrl
-            caption
-            }
-            }
-            id
-            name
-            categoryValues {
-            id
-            name
-            category {
-            id
-            name
-            itemType
-            }
-            }
-            recipeInstructions {
-            text
-            position
-            }
-            recipeTreeComponents(levels: [1]) {
-            quantityUnitValues {
-            value
-            unit {
-            id
-            name
-            }
-            }
-            recipeItem {
-            preparations {
-            id
-            name
-            }
-            subRecipe {
-            id
-            name
-            externalName
-            recipeTreeComponents(levels: [1]) {
-            quantityUnitValues {
-            value
-            unit {
-            id
-            name
-            }
-            }
-            ingredient {
-            id
-            name
-            externalName
-            }
-            recipeItem {
-            preparations {
-            id
-            name
-            }
-            subRecipe {
-            id
-            name
-            externalName
-            recipeInstructions {
-            text
-            position
-            }
-            }
-            }
-            }
-            }
-            }
-            }
-            }
-            volume
-            }
-            }
-            }
-            }'''.replace(' '*12, '')
-            
-    def test_get_ops_menu_query(self):
-        query = get_ops_menu_query(dates=["2022-03-28"])
-        query_str = bytes(query).decode('utf-8')
-        self.maxDiff = None
-        self.assertEqual(query_str, self.expected_query)
